Fix an error with scoping in my function
I thought that it was a lexical scoping issue but it didn't appear to
be. My file has lexical scoping turned on at the top but I think the
error was a lexical scoping one.

I added a new private directory with some useful error handling during
startup to indicate an error.